Configure: fixed --with-libatomic=DIR with recent libatomic_ops.
The build location of the resulting libatomic_ops.a was changed in v7.4.0
after converting libatomic_ops to use libtool.  The fix is to use headers
and library from the install prefix, similar how we do this with OpenSSL.
This approach allows building with both old and new library versions.

Initially reported here:
https://mailman.nginx.org/pipermail/nginx/2018-April/056054.html